% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\docType{methods}
\name{apply_transformations,VEST-method}
\alias{apply_transformations,VEST-method}
\title{Apply transformations using VEST}
\usage{
\S4method{apply_transformations}{VEST}(object)
}
\arguments{
\item{object}{object of class VEST}
}
\description{
Apply transformations using VEST
}
