% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\name{apply_transformations}
\alias{apply_transformations}
\title{Apply transformations generic}
\usage{
apply_transformations(object)
}
\arguments{
\item{object}{object of class VEST}
}
\description{
Apply transformations generic
}
\keyword{internal}
