% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{BayesianSignTest}
\alias{BayesianSignTest}
\title{Bayes Sign Test}
\usage{
BayesianSignTest(diffVector, rope_min, rope_max)
}
\arguments{
\item{diffVector}{diffVector}

\item{rope_min}{rope_min}

\item{rope_max}{rope_min}
}
\description{
Bayes Sign Test
}
