% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{create_datasets}
\alias{create_datasets}
\title{Create train/test data sets}
\usage{
create_datasets(train, test, k, h)
}
\arguments{
\item{train}{train series, vector}

\item{test}{test series, vector, posterior to train}

\item{k}{embedding dimension}

\item{h}{forecasting horizon}
}
\description{
Create train/test data sets
}
