% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-step-forecast.r
\name{direct_msf}
\alias{direct_msf}
\title{Direct multi-step forecasting}
\usage{
direct_msf(X_tr, X_ts_lags, X_ts_dynamics, y_tr, h, training_fun, predict_fun)
}
\arguments{
\item{X_tr}{assad}

\item{X_ts_lags}{sss}

\item{X_ts_dynamics}{as}

\item{y_tr}{as}

\item{h}{ss}

\item{training_fun}{sas}

\item{predict_fun}{ss}
}
\description{
Direct multi-step forecasting
}
