% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-expsmooth.r
\name{ets_multistep}
\alias{ets_multistep}
\title{ETS MODEL}
\usage{
ets_multistep(train, test, freq, h)
}
\arguments{
\item{train}{train series, ts}

\item{test}{test series, ts}

\item{freq}{frequency}

\item{h}{horizon}
}
\description{
ETS MODEL
}
