% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-selection.r
\name{feat_select_bf}
\alias{feat_select_bf}
\title{Feature Selection, best summary by rep2}
\usage{
feat_select_bf(x)
}
\arguments{
\item{x}{object of class VEST}
}
\description{
Feature Selection, best summary by rep2
}
