% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-selection.r
\name{feat_select_corr}
\alias{feat_select_corr}
\title{Feature Selection, best summary by rep3}
\usage{
feat_select_corr(x)
}
\arguments{
\item{x}{object of class VEST}
}
\description{
Feature Selection, best summary by rep3
}
