% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-selection.r
\name{feat_select_vest}
\alias{feat_select_vest}
\title{Feature Selection, best summary by rep}
\usage{
feat_select_vest(x)
}
\arguments{
\item{x}{object of class VEST}
}
\description{
Feature Selection, best summary by rep
}
