% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\name{feature_engineering}
\alias{feature_engineering}
\title{Feature Engineering Wrapper}
\usage{
feature_engineering(x, targets, freq, compute_importance = TRUE)
}
\arguments{
\item{x}{Embedded time series}

\item{targets}{target column names}

\item{freq}{frequency of the time series}

\item{compute_importance}{compute_importance}
}
\description{
Feature Engineering Wrapper
}
