% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-transformations.r
\name{ft_t_fourier_cos}
\alias{ft_t_fourier_cos}
\title{First Differsssaassadences transf.}
\usage{
ft_t_fourier_cos(x, freq, past_l)
}
\arguments{
\item{x}{df}
}
\description{
First Differsssaassadences transf.
}
