% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-transformations.r
\name{ft_t_fourier_sin}
\alias{ft_t_fourier_sin}
\title{First Differences transf.asdada}
\usage{
ft_t_fourier_sin(x, freq, past_l)
}
\arguments{
\item{x}{df}
}
\description{
First Differences transf.asdada
}
