% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-transformations.r
\name{ft_t_sdiff}
\alias{ft_t_sdiff}
\title{Seasonal differencing}
\usage{
ft_t_sdiff(x, freq, last_freq_x = NULL)
}
\arguments{
\item{x}{df}

\item{freq}{freq}

\item{last_freq_x}{last_freq_x}
}
\description{
Seasonal differencing
}
