% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-transformations.r
\name{ft_t_seasadj}
\alias{ft_t_seasadj}
\title{Seasonal adjustment}
\usage{
ft_t_seasadj(x, freq, past_x = NULL)
}
\arguments{
\item{x}{df}

\item{freq}{freq}

\item{past_x}{last_freq_x}
}
\description{
Seasonal adjustment
}
