% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-transformations.r
\name{ft_transformation}
\alias{ft_transformation}
\title{Feature transformations}
\usage{
ft_transformation(x, freq)
}
\arguments{
\item{x}{Predictor variables of embedded time series}

\item{freq}{frequency of time series}
}
\description{
Feature transformations
}
