% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\name{get_fourier_terms}
\alias{get_fourier_terms}
\title{Dynamic Harmonic Regre. Terms}
\usage{
get_fourier_terms(object, nterms = 3)
}
\arguments{
\item{object}{object of class VEST}

\item{nterms}{no terms. def to 3}
}
\description{
Dynamic Harmonic Regre. Terms
}
