% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-summary.r
\name{get_fourier_terms_}
\alias{get_fourier_terms_}
\title{Get fourieer terms for test}
\usage{
get_fourier_terms_(y, nterms, freq, h)
}
\arguments{
\item{y}{time series}

\item{nterms}{nterms}

\item{freq}{frequency}

\item{h}{horizon}
}
\description{
Get fourieer terms for test
}
\keyword{internal}
