% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-summary.r
\name{get_importance,VEST-method}
\alias{get_importance,VEST-method}
\title{Get importance scores}
\usage{
\S4method{get_importance}{VEST}(object)
}
\arguments{
\item{object}{object of class VEST}
}
\description{
Get importance scores
}
