% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\name{get_outer_dynamics}
\alias{get_outer_dynamics}
\title{Get outer dynamics
DHR}
\usage{
get_outer_dynamics(object)
}
\arguments{
\item{object}{object of class VEST}
}
\description{
Get outer dynamics
DHR
}
