% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-summary.r
\docType{methods}
\name{get_summary,VEST-method}
\alias{get_summary,VEST-method}
\title{Get summary operations}
\usage{
\S4method{get_summary}{VEST}(object)
}
\arguments{
\item{object}{object of class VEST}
}
\description{
Get summary operations
}
