% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\docType{methods}
\name{get_transformations,VEST-method}
\alias{get_transformations,VEST-method}
\title{Transform the original representation}
\usage{
\S4method{get_transformations}{VEST}(object)
}
\arguments{
\item{object}{object of class VEST}
}
\description{
Transform the original representation
}
\keyword{internal}
