% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\name{get_transformations}
\alias{get_transformations}
\title{Get transformations generic fun}
\usage{
get_transformations(object)
}
\arguments{
\item{object}{object of class VEST}
}
\description{
Get transformations generic fun
}
\keyword{internal}
