% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.r
\name{get_xgb_model}
\alias{get_xgb_model}
\title{Create xgb model using a grid search}
\usage{
get_xgb_model(form, data)
}
\arguments{
\item{form}{formula}

\item{data}{training data, data.frame}
}
\description{
Create xgb model using a grid search
}
