% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-algorithms.r
\name{M5.cycle}
\alias{M5.cycle}
\title{wrappersadda}
\usage{
M5.cycle(form, train, test, ...)
}
\arguments{
\item{form}{form}

\item{train}{sadasda}

\item{test}{sasa}

\item{...}{sadada}
}
\description{
wrappersadda
}
