% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-algorithms.r
\name{M5.multi_step_cycle}
\alias{M5.multi_step_cycle}
\title{Wrapper M5 multistep cycle}
\usage{
M5.multi_step_cycle(train, test, h)
}
\arguments{
\item{train}{train set, data.frame}

\item{test}{test set, data.frame}

\item{h}{forecasting horizon}
}
\description{
Wrapper M5 multistep cycle
}
