% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-algorithms.r
\name{M5.predict}
\alias{M5.predict}
\title{Model tree prediction function}
\usage{
M5.predict(model, test)
}
\arguments{
\item{model}{model from M5.train}

\item{test}{test, data.frame}
}
\description{
Model tree prediction function
}
