% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-algorithms.r
\name{M5.self_cycle}
\alias{M5.self_cycle}
\title{Wrapper for M5 cycle using training data}
\usage{
M5.self_cycle(form, train, ...)
}
\arguments{
\item{form}{formula}

\item{train}{train set, data.frame}

\item{...}{other pars, to M5.cycle}
}
\description{
Wrapper for M5 cycle using training data
}
