% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-algorithms.r
\name{M5.train}
\alias{M5.train}
\title{Model tree using cubist}
\usage{
M5.train(form, train, iterations = 25)
}
\arguments{
\item{form}{formula}

\item{train}{train set, data.frame}

\item{iterations}{no. boosting iterations}
}
\description{
Model tree using cubist
}
