% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-step-forecast.r
\name{multistep.prediction}
\alias{multistep.prediction}
\title{Multi step prediction with updated dynamics}
\usage{
multistep.prediction(model, x, dyns, h, ft_model, update.dynamics = TRUE,
  model_cols, predfun)
}
\arguments{
\item{model}{Predictive model}

\item{x}{Past lags in the test data}

\item{dyns}{Dynamics in the test data (for h=1)}

\item{h}{Forecasting horizon}

\item{ft_model}{Feat engineer model for updating dynamics}

\item{update.dynamics}{logical. If true, dynamics are updated for h>1}

\item{model_cols}{Names of the features in trained model}

\item{predfun}{Prediction fun for making predictions using model}
}
\description{
Multi step prediction with updated dynamics
}
