% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-delay-embedding.r
\name{my_embedd}
\alias{my_embedd}
\title{Embedding function}
\usage{
my_embedd(x, m, d = 1, h = 1)
}
\arguments{
\item{x}{time series, vector}

\item{m}{embedding dimension, int}

\item{d}{delay parameter, def to 1}

\item{h}{forecasting horizon, def to 1}
}
\description{
Embedding function
}
