% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\name{pred_fourier_terms}
\alias{pred_fourier_terms}
\title{Predicting new DHR terms}
\usage{
pred_fourier_terms(object, nterms = 3, h)
}
\arguments{
\item{object}{object of class VEST}

\item{nterms}{no of terms}

\item{h}{forecasting horizon}
}
\description{
Predicting new DHR terms
}
