% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\name{predict_outer_dynamics}
\alias{predict_outer_dynamics}
\title{Predicting new outer dynamics}
\usage{
predict_outer_dynamics(object, test)
}
\arguments{
\item{object}{object of class VEST}

\item{test}{new observations}
}
\description{
Predicting new outer dynamics
}
