% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{rep_holdout_origins}
\alias{rep_holdout_origins}
\title{rep h origins}
\usage{
rep_holdout_origins(n, nreps, train_size, test_size)
}
\arguments{
\item{n}{len}

\item{nreps}{n sims}

\item{train_size}{ratio of len}

\item{test_size}{ratio of len}
}
\description{
rep h origins
}
