% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-expsmooth.r
\name{run_classical_methods}
\alias{run_classical_methods}
\title{run clasical methds}
\usage{
run_classical_methods(y_train, y_test, freq, h)
}
\arguments{
\item{y_train}{num}

\item{y_test}{num}

\item{freq}{frq}

\item{h}{h}
}
\description{
run clasical methds
}
