% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-delay-embedding.r
\name{search_k_vall}
\alias{search_k_vall}
\title{Search best embedding dimension
using model tree performance}
\usage{
search_k_vall(x, max_k = 30)
}
\arguments{
\item{x}{time series, vector}

\item{max_k}{maximum embedding dimension. Def to 30}
}
\description{
Search best embedding dimension
using model tree performance
}
