% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-expsmooth.r
\name{tbats_multistep}
\alias{tbats_multistep}
\title{TBATS MODEL}
\usage{
tbats_multistep(train, test, freq, h)
}
\arguments{
\item{train}{train series, ts}

\item{test}{test series, ts}

\item{freq}{frequency}

\item{h}{horizon}
}
\description{
TBATS MODEL
}
