% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ts_df_slip_by_size}
\alias{ts_df_slip_by_size}
\title{Splitting data}
\usage{
ts_df_slip_by_size(x, test_size)
}
\arguments{
\item{x}{dataset, data.frame}

\item{test_size}{test size, no of obs.}
}
\description{
Splitting data
}
