% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ts_holdout}
\alias{ts_holdout}
\title{Holdout estimation}
\usage{
ts_holdout(x, ratio, frq)
}
\arguments{
\item{x}{time series, ts class}

\item{ratio}{ratio for training}

\item{frq}{frequency of time series}
}
\description{
Holdout estimation
}
