% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{unroll_embedd}
\alias{unroll_embedd}
\title{Unroll embedded time series}
\usage{
unroll_embedd(df)
}
\arguments{
\item{df}{time series, data.frame (my_embedd fun)}
}
\description{
Unroll embedded time series
}
