% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\docType{class}
\name{VEST-class}
\alias{VEST-class}
\title{Class for time series feature eng.}
\description{
Class for time series feature eng.
}
\section{Slots}{

\describe{
\item{\code{X}}{Predictor variables. Past recent values}

\item{\code{y}}{target values. numeric if h=1, df if h>1}

\item{\code{Z}}{list of representations of X}

\item{\code{Dynamics}}{statistics derived from Z and X}

\item{\code{keys}}{metadata}

\item{\code{importance}}{importance scores of features}
}}

