% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\name{VEST}
\alias{VEST}
\title{Data.frame constructor for VEST class}
\usage{
VEST(targets, x)
}
\arguments{
\item{targets}{character vector describing the target variable(s)}

\item{x}{embedded time series as data.frame}
}
\description{
Data.frame constructor for VEST class
}
