% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.r
\name{WF_part1}
\alias{WF_part1}
\title{WF part 1
Getting the VEST features}
\usage{
WF_part1(train, test, h)
}
\arguments{
\item{train}{time series, ts class}

\item{test}{test time series, ts class}

\item{h}{forecasting horizon}
}
\description{
WF part 1
Getting the VEST features
}
