% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.r
\name{WF_part2}
\alias{WF_part2}
\title{WF part2}
\usage{
WF_part2(x, learning_functions)
}
\arguments{
\item{x}{results from part1}

\item{learning_functions}{list with 2 elements.
the frist element is the training func, and
the second element is the predict func}
}
\description{
WF part2
}
