% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.r
\name{WF_part2_direct2}
\alias{WF_part2_direct2}
\title{Workflow part 2
Training forecasting models
MSF with a direct approach}
\usage{
WF_part2_direct2(x, h, learning_functions)
}
\arguments{
\item{x}{results from wf pt1}

\item{h}{horizon}

\item{learning_function}{learning functions.
A list: first element is the training function, second element is the predict funcions}
}
\description{
Workflow part 2
Training forecasting models
MSF with a direct approach
}
