% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-algorithms.r
\name{XGB.predict}
\alias{XGB.predict}
\title{XGB predict function}
\usage{
XGB.predict(model, test)
}
\arguments{
\item{model}{model from XGB.train fun}

\item{test}{test set, data.frame}
}
\description{
XGB predict function
}
