% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-algorithms.r
\name{XGB.train}
\alias{XGB.train}
\title{XGB training function}
\usage{
XGB.train(form, train)
}
\arguments{
\item{form}{formula}

\item{train}{train set, data.frame}
}
\description{
XGB training function
}
