% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.r
\name{arima_multistep}
\alias{arima_multistep}
\title{arima baseline}
\usage{
arima_multistep(train, test, freq, h)
}
\arguments{
\item{train}{train series, ts}

\item{test}{test series, ts}

\item{freq}{frequency}

\item{h}{horizon}
}
\description{
arima baseline
}
