% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-summary.r
\docType{methods}
\name{cleanup_feats,VEST-method}
\alias{cleanup_feats,VEST-method}
\title{Feature clean up method}
\usage{
\S4method{cleanup_feats}{VEST}(object)
}
\arguments{
\item{object}{object of class VEST}
}
\description{
Feature clean up method
}
