% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-algorithms.r
\name{M5.cycle}
\alias{M5.cycle}
\title{Wrapper for M5 model}
\usage{
M5.cycle(form, train, test, ...)
}
\arguments{
\item{form}{formula}

\item{train}{train set, data.frame}

\item{test}{test set, data.frame}

\item{...}{other parameters to m5.train}
}
\description{
Wrapper for M5 model
}
