% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-algorithms.r
\name{M5.self_cycle}
\alias{M5.self_cycle}
\title{wrapperq}
\usage{
M5.self_cycle(form, train, ...)
}
\arguments{
\item{form}{forms}

\item{train}{sadasdsa}

\item{...}{sadadas}
}
\description{
wrapperq
}
