% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vest.r
\name{predict,VEST-method}
\alias{predict,VEST-method}
\title{Predict feature values with new observations}
\usage{
\S4method{predict}{VEST}(object, x, update_last_rows = FALSE)
}
\arguments{
\item{object}{object of class VEST}

\item{x}{x past lags w/o targets}
}
\description{
Predict feature values with new observations
}
